% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records_by_progetti.R
\name{records_by_progetti}
\alias{records_by_progetti}
\title{Get the data}
\usage{
records_by_progetti(API_Key = FALSE, territorio = NULL, tema = NULL,
  natura = NULL)
}
\arguments{
\item{API_Key}{API Key present}

\item{territorio}{Territorio}

\item{tema}{Tema}

\item{natura}{Natura}
}
\description{
Get the data from the API
}

