% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropencoesione-package.R
\docType{package}
\name{ropencoesione-package}
\alias{ropencoesione}
\alias{ropencoesione-package}
\title{OpenCoesione data}
\description{
OpenCoesione data
}
\author{
Patrick Hausmann \email{patrick.hausmann@covimo.de}
}
\keyword{package}

